requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.5
pycparser==2.21
PyNaCl==1.5.0
redis==4.5.4
requests==2.29.0
six==1.16.0
spotipy==2.23.0
tqdm==4.65.0
urllib3==1.26.15
yarl==1.9.2
